Mon 1/2/2021 12am

+++ GroupA/JavaFundamentals/src/ie/gmit/Lesson01/Activity02/ReadScanner.java
@@ -1,6 +1,5 @@
 package ie.gmit.Lesson01.Activity02;
 
-
 import java.util.Scanner;
 
 public class ReadScanner {
@@ -8,9 +7,9 @@ public class ReadScanner {
   public static void main(String[] args) {
     Scanner sc = new Scanner(System.in);
     System.out.print("Enter a number: ");
-    double a = sc.nextDouble();
+    int a = sc.nextInt();
     System.out.print("Enter 2nd number: ");
-    double b = sc.nextDouble();
+    int b = sc.nextInt();
     System.out.println("The sum is " + (a + b) + ".");
 
   }
