+++ GroupB/JavaFundamentals/src/ie/gmit/Lesson01/Activity02/ReadScanner.java
@@ -1,5 +1,6 @@
 package ie.gmit.Lesson01.Activity02;
 
+
 import java.util.Scanner;
 
 public class ReadScanner {
@@ -7,9 +8,9 @@ public class ReadScanner {
   public static void main(String[] args) {
     Scanner sc = new Scanner(System.in);
     System.out.print("Enter a number: ");
-    int a = sc.nextInt();
+    double a = sc.nextDouble();
     System.out.print("Enter 2nd number: ");
-    int b = sc.nextInt();
+    double b = sc.nextDouble();
     System.out.println("The sum is " + (a + b) + ".");
 
   }
